factor out the structured outputs into the LLMWrapper
