+++ utils.js
@@ -1,4 +1,6 @@
 import OpenAI from "openai";
+import { z } from "zod";
+import { zodResponseFormat } from "openai/helpers/zod";
 
 let utils = {};
 
@@ -143,7 +145,7 @@ export class LLMWrapper {
   #openAIKey;
   #googleKey;
   
-  model = new ModelCapabilities('gpt-4o');
+  model = new ModelCapabilities(LLMWrapper.DEFAULT_MODEL);
   openAIAPI = null;
 
   constructor(parameters) {
@@ -182,6 +184,40 @@ export class LLMWrapper {
             });
             break;
     }
+  }
+
+  static SCHEMA_STRINGS = {
+    "from": "This is a variable which causes the to variable in this relationship that is between two variables, from and to.  The from variable is the equivalent of a cause.  The to variable is the equivalent of an effect",
+    "to": "This is a variable which is impacted by the from variable in this relationship that is between two variables, from and to.  The from variable is the equivalent of a cause.  The to variable is the equivalent of an effect",
+    "reasoning": "This is an explanation for why this relationship exists",
+    "polarity": "There are two possible kinds of relationships.  The first are relationships with positive polarity that are represented with a + symbol.  In relationships with positive polarity (+) a change in the from variable causes a change in the same direction in the to variable.  For example, in a relationship with postive polarity (+), a decrease in the from variable, would lead to a decrease in the to variable.  The second kind of relationship are those with negative polarity that are represented with a - symbol.  In relationships with negative polarity (-) a change in the from variable causes a change in the opposite direction in the to variable.  For example, in a relationship with negative polarity (-) an increase in the from variable, would lead to a decrease in the to variable.",
+    "polarityReasoning": "This is the reason for why the polarity for this relationship was choosen",
+    "relationship": "This is a relationship between two variables, from and to (from is the cause, to is the effect).  The relationship also contains a polarity which describes how a change in the from variable impacts the to variable",
+    "relationships": "The list of relationships you think are appropriate to satisfy my request based on all of the information I have given you",
+    "explanation": "Concisely explain your reasoning for each change you made to the old CLD to create the new CLD. Speak in plain English, don't reference json specifically. Don't reiterate the request or any of these instructions.",
+    "title": "A highly descriptive 7 word max title describing your explanation."
+};
+
+  static DEFAULT_MODEL = 'gpt-4o';
+
+  generateResponseSchema() {
+      const PolarityEnum = z.enum(["+", "-"]).describe(LLMWrapper.SCHEMA_STRINGS.polarity);
+
+      const Relationship = z.object({
+          from: z.string().describe(LLMWrapper.SCHEMA_STRINGS.from),
+          to: z.string().describe(LLMWrapper.SCHEMA_STRINGS.to),
+          polarity: PolarityEnum,
+          reasoning: z.string().describe(LLMWrapper.SCHEMA_STRINGS.reasoning),
+          polarityReasoning: z.string().describe(LLMWrapper.SCHEMA_STRINGS.polarityReasoning)
+      }).describe(LLMWrapper.SCHEMA_STRINGS.relationship);
+          
+      const Relationships = z.object({
+          explanation: z.string().describe(LLMWrapper.SCHEMA_STRINGS.explanation),
+          title: z.string().describe(LLMWrapper.SCHEMA_STRINGS.title),
+          relationships: z.array(Relationship).describe(LLMWrapper.SCHEMA_STRINGS.relationships)
+      });
+
+      return zodResponseFormat(Relationships, "relationships_response");
   }
 
   static additionalParameters() {
@@ -223,7 +259,7 @@ export class LLMWrapper {
         },{
             name: "underlyingModel",
             type: "string",
-            defaultValue: 'gpt-4o',
+            defaultValue: LLMWrapper.DEFAULT_MODEL,
             required: false,
             options: models,
             uiElement: "combobox",
